+++ packages/tsurlfilter/src/modifiers/domain-modifier.ts
@@ -227,7 +227,7 @@ export class DomainModifier {
                         return true;
                     }
                 } catch {
-                    logger.error(`Invalid regular expression as domain pattern: "${d}"`);
+                    logger.error(`[tsurl.DomainModifier.isDomainOrSubdomainOfAny]: invalid regular expression as domain pattern: "${d}"`);
                 }
                 continue;
             }