+++ packages/tswebextension/src/lib/mv3/background/cosmetic-api.ts
@@ -1,12 +1,11 @@
 import { type ScriptletData, type CosmeticResult, type CosmeticRule } from '@adguard/tsurlfilter';
 
 import { CosmeticApiCommon, type ContentScriptCosmeticData, type LogJsRulesParams } from '../../common/cosmetic-api';
-import { getErrorMessage } from '../../common/error';
 import { createFrameMatchQuery } from '../../common/utils/create-frame-match-query';
 import { logger } from '../../common/utils/logger';
-import { getDomain } from '../../common/utils/url';
+import { getDomain, isExtensionUrl } from '../../common/utils/url';
 import { tabsApi } from '../tabs/tabs-api';
-import { USER_FILTER_ID } from '../../common/constants';
+import { BACKGROUND_TAB_ID, USER_FILTER_ID } from '../../common/constants';
 
 import { appContext } from './app-context';
 import { engineApi } from './engine-api';
@@ -221,7 +220,7 @@ export class CosmeticApi extends CosmeticApiCommon {
                 });
             }));
         } catch (e) {
-            logger.debug('[applyJsFuncsByTabAndFrame] error occurred during injection', getErrorMessage(e));
+            logger.info('[tsweb.CosmeticApi.applyJsFuncsByTabAndFrame]: error occurred during injection: ', e);
         }
     }
 
@@ -255,9 +254,26 @@ export class CosmeticApi extends CosmeticApiCommon {
                 });
             }));
         } catch (e) {
-            // TODO: getErrorMessage may not be needed since logger should handle arg types
-            logger.debug('[applyScriptletsByTabAndFrame] error occurred during injection', getErrorMessage(e));
+            logger.info('[tsweb.CosmeticApi.applyScriptletsByTabAndFrame]: error occurred during injection: ', e);
         }
+    }
+
+    /**
+     * Checks if cosmetics should be applied —
+     * background page or extension pages do not need cosmetics applied.
+     *
+     * @param tabId Tab id.
+     * @param frameUrl Frame url.
+     *
+     * @returns True if cosmetics should be applied, false otherwise.
+     */
+    public static shouldApplyCosmetics(tabId: number, frameUrl: string): boolean {
+        // no need to apply cosmetic rules on background or extension pages
+        if (tabId === BACKGROUND_TAB_ID || isExtensionUrl(frameUrl)) {
+            return false;
+        }
+
+        return true;
     }
 
     /**
@@ -281,12 +297,7 @@ export class CosmeticApi extends CosmeticApiCommon {
                 frameId,
             });
         } catch (e) {
-            logger.debug(
-                '[applyCssByTabAndFrame] error occurred during injection',
-                getErrorMessage(e),
-                'with frame context:',
-                frameContext,
-            );
+            logger.info('[tsweb.CosmeticApi.applyCssByTabAndFrame]: error occurred during injection: ', e, 'with frame context:', frameContext);
         }
     }
 
@@ -319,10 +330,7 @@ export class CosmeticApi extends CosmeticApiCommon {
                 scriptText,
             });
         } catch (e) {
-            logger.debug(
-                '[applyJsFuncsAndScriptletsByTabAndFrame] error occurred during injection',
-                getErrorMessage(e),
-            );
+            logger.info('[tsweb.CosmeticApi.applyJsFuncsAndScriptletsByTabAndFrame]: error occurred during injection: ', e);
         }
     }