+++ packages/tswebextension/src/lib/mv3/tabs/tabs-cosmetic-injector.ts
@@ -30,7 +30,7 @@ export class TabsCosmeticInjector {
         // Handles errors
         promises.forEach((promise) => {
             if (promise.status === 'rejected') {
-                logger.error('[tswebextension.processOpenTabs]: cannot inject cosmetic to open tab: ', promise.reason);
+                logger.error('[tsweb.TabsCosmeticInjector.processOpenTabs]: cannot inject cosmetic to open tab: ', promise.reason);
             }
         });
 
@@ -102,6 +102,11 @@ export class TabsCosmeticInjector {
                 return;
             }
 
+            if (!CosmeticApi.shouldApplyCosmetics(tabId, url)) {
+                logger.debug(`[tsweb.TabsCosmeticInjector.processOpenTab]: skipping cosmetics injection for background or extension page with tabId ${tabId}, frameId ${frameId} and url ${url}`);
+                return;
+            }
+
             const tasks = [
                 CosmeticApi.applyCssByTabAndFrame(tabId, frameId),
             ];
@@ -117,12 +122,14 @@ export class TabsCosmeticInjector {
             // since it is used not only here.
             // Note: this is an async function, but we will not await it because
             // events do not support async listeners.
-            Promise.all(tasks).catch((e) => logger.error(e));
+            Promise.all(tasks).catch((e) => {
+                logger.error('[tsweb.TabsCosmeticInjector.processOpenTab]: cannot apply cosmetic rules: ', e);
+            });
 
             const frameContext = tabsApi.getFrameContext(tabId, frameId);
             if (!frameContext?.cosmeticResult) {
                 // eslint-disable-next-line max-len
-                logger.debug(`[tswebextension.processOpenTab]: cannot log script rules due to not having cosmetic result for tabId: ${tabId}, frameId: ${frameId}.`);
+                logger.debug(`[tsweb.TabsCosmeticInjector.processOpenTab]: cannot log script rules due to not having cosmetic result for tabId: ${tabId}, frameId: ${frameId}.`);
                 return;
             }