+++ packages/dnr-rulesets/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `watch` command to watch for changes in the filters
   and automatically rebuild DNR.
 
+### Changed
+
+- Updated [@adguard/logger] to `v2.0.0`.
+
 ## [3.0.0-alpha.12] - 2025-05-28
 
 ### Changed
@@ -128,3 +132,5 @@ For example, while bundling the manifest with the extension.
 ### Changed
 
 - Improve filters compatibility with Declarative Net Request API.
+
+[@adguard/logger]: ../logger/CHANGELOG.md