+++ packages/tswebextension/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `filterId` query param to `documentBlockingPageUrl` for MV2.
+- Extension blocking page support for requests blocked by `$document` rules in MV3.
+- `filterId` query param to `documentBlockingPageUrl` for both MV2 and MV3.
 - Support for Chrome's User Scripts API that allows more reliable script
   injection in MV3 if developer mode is enabled.
 - Static getter `isUserScriptsApiSupported` to check if the current browser
@@ -40,6 +41,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [AdguardBrowserExtension#3122]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3122
 [AdguardBrowserExtension#3164]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3164
 [AdguardBrowserExtension#3192]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3192
+
+### Changed
+
+- Updated [@adguard/logger] to `v2.0.0`.
 
 ## [3.1.0] - 2025-05-28